orm/model/model_test.go
package model

import (
	"database/sql"
	"gitee.com/geektime-geekbang/geektime-go/orm/internal/errs"
	"github.com/stretchr/testify/assert"
	"github.com/stretchr/testify/require"
	"reflect"
	"testing"
)

func Test_registry_Register(t *testing.T) {
	testCases := []struct{
		name string
		entity any
		wantModel *Model
		wantErr error
	}{
		{
			name:    "struct",
			entity:  TestModel{},
			wantErr: errs.ErrPointerOnly,
		},
		{
			name: "pointer",
			entity: &TestModel{},
			wantModel: &Model{
				TableName: "test_model",
				Fields: []*Field{
					{
						ColName: "id",
						GoName:  "Id",
						Type:    reflect.TypeOf(int64(0)),
					},
					{
						ColName: "first_name",
						GoName:  "FirstName",
						Type:    reflect.TypeOf(""),
						Offset: 8,
					},
					{
						ColName: "age",
						GoName:  "Age",
						Type:    reflect.TypeOf(int8(0)),
						Offset: 24,
					},
					{
						ColName: "last_name",
						GoName:  "LastName",
						Type:    reflect.TypeOf(&sql.NullString{}),
						Offset: 32,
					},
				},
			},
		},
		{
			name: "map",
			entity: map[string]string{},
			wantErr: errs.ErrPointerOnly,
		},
		{
			name: "slice",
			entity: []int{},
			wantErr: errs.ErrPointerOnly,
		},
		{
			name: "basic types",
			entity: 0,
			wantErr: errs.ErrPointerOnly,
		},
	}

	r := &registry{}
	for _, tc := range testCases {
		t.Run(tc.name, func(t *testing.T) {
			m, err := r.Register(tc.entity)
			assert.Equal(t, tc.wantErr, err)
			if err != nil {
				return
			}
			fieldMap := make(map[string]*Field)
			columnMap := make(map[string]*Field)
			for _, f := range tc.wantModel.Fields {
				fieldMap[f.GoName] = f
				columnMap[f.ColName] = f
			}
			tc.wantModel.FieldMap = fieldMap
			tc.wantModel.ColumnMap = columnMap
			assert.Equal(t, tc.wantModel, m)
		})
	}
}

func TestRegistry_get(t *testing.T) {
	testCases := []struct{
		name string

		entity any
		wantModel *Model
		wantErr error
	} {
		{
			name: "pointer",
			entity: &TestModel{},
			wantModel: &Model{
				TableName: "test_model",
				Fields:  []*Field{
					{
						ColName: "id",
						GoName:  "Id",
						Type:    reflect.TypeOf(int64(0)),
					},
					{
						ColName: "first_name",
						GoName:  "FirstName",
						Type:    reflect.TypeOf(""),
						Offset: 8,
					},
					{
						ColName: "age",
						GoName:  "Age",
						Type:    reflect.TypeOf(int8(0)),
						Offset: 24,
					},
					{
						ColName: "last_name",
						GoName:  "LastName",
						Type:    reflect.TypeOf(&sql.NullString{}),
						Offset: 32,
					},
				},
			},
		},
		{
			name: "tag",
			entity: func() any {
				type TagTable struct {
					FirstName string `orm:"column=first_name_t"`
				}
				return &TagTable{}
			}(),
			wantModel: &Model{
				TableName: "tag_table",
				Fields: []*Field{
					{
						ColName: "first_name_t",
						GoName:  "FirstName",
						Type:    reflect.TypeOf(""),
					},
				},
			},
		},
		{
			name: "empty column",
			entity: func() any {
				type TagTable struct {
					FirstName string `orm:"column="`
				}
				return &TagTable{}
			}(),
			wantModel: &Model{
				TableName: "tag_table",
				Fields: []*Field{
					{
						ColName: "first_name",
						GoName:  "FirstName",
						Type:    reflect.TypeOf(""),
					},
				},
			},
		},
		{
			name: "column only",
			entity: func() any {
				type TagTable struct {
					FirstName string `orm:"column"`
				}
				return &TagTable{}
			}(),
			wantErr: errs.NewErrInvalidTagContent("column"),
		},
		{
			name: "ignore tag",
			entity: func() any {
				type TagTable struct {
					FirstName string `orm:"abc=abc"`
				}
				return &TagTable{}
			}(),
			wantModel: &Model{
				TableName: "tag_table",
				Fields: []*Field{
					{
						ColName: "first_name",
						GoName:  "FirstName",
						Type:    reflect.TypeOf(""),
					},
				},
			},
		},
		{
			name: "table name",
			entity: &CustomTableName{},
			wantModel: &Model{
				TableName: "custom_table_name_t",
				Fields: []*Field{
					{
						ColName: "first_name",
						GoName:  "FirstName",
						Type:    reflect.TypeOf(""),
					},
				},
			},
		},
		{
			name: "table name ptr",
			entity: &CustomTableNamePtr{},
			wantModel: &Model{
				TableName: "custom_table_name_ptr_t",
				Fields: []*Field{
					{
						ColName: "first_name",
						GoName:  "FirstName",
						Type:    reflect.TypeOf(""),
					},
				},
			},
		},
		{
			name: "empty table name",
			entity: &EmptyTableName{},
			wantModel: &Model{
				TableName: "empty_table_name",
				Fields: []*Field{
					{
						ColName: "first_name",
						GoName:  "FirstName",
						Type:    reflect.TypeOf(""),
					},
				},
			},
		},
	}
	r := NewRegistry()
	for _, tc := range testCases {
		t.Run(tc.name, func(t *testing.T) {
			m, err := r.Get(tc.entity)
			assert.Equal(t, tc.wantErr, err)
			if err != nil {
				return
			}

			fieldMap := make(map[string]*Field)
			columnMap := make(map[string]*Field)
			for _, f := range tc.wantModel.Fields {
				fieldMap[f.GoName] = f
				columnMap[f.ColName] = f
			}
			tc.wantModel.FieldMap = fieldMap
			tc.wantModel.ColumnMap = columnMap

			assert.Equal(t, tc.wantModel, m)

			typ := reflect.TypeOf(tc.entity)
			cache, ok := r.(*registry).models.Load(typ)
			assert.True(t, ok)
			assert.Equal(t, tc.wantModel, cache)
		})
	}
}

type CustomTableName struct {
	FirstName string
}

func (c CustomTableName) TableName() string {
	return "custom_table_name_t"
}


type CustomTableNamePtr struct {
	FirstName string
}

func (c *CustomTableNamePtr) TableName() string {
	return "custom_table_name_ptr_t"
}

type EmptyTableName struct {
	FirstName string
}

func (c *EmptyTableName) TableName() string {
	return ""
}

func TestModelWithTableName(t *testing.T) {
	r := NewRegistry()
	m, err := r.Register(&TestModel{}, WithTableName("test_model_ttt"))
	require.NoError(t, err)
	assert.Equal(t, "test_model_ttt", m.TableName)
}

func TestModelWithColumnName(t *testing.T) {
	testCases := []struct {
		name string
		field string
		colName string

		wantColName string
		wantErr error
	}{
		{
			name: "column name",
			field: "FirstName",
			colName: "first_name_ccc",
			wantColName: "first_name_ccc",
		},
		{
			name: "invalid column name",
			field: "XXX",
			colName: "first_name_ccc",
			wantErr: errs.NewErrUnknownField("XXX"),
		},
	}

	for _, tc := range testCases {
		t.Run(tc.name, func(t *testing.T) {
			r := NewRegistry()
			m, err := r.Register(&TestModel{}, WithColumnName(tc.field, tc.colName))
			assert.Equal(t, tc.wantErr, err)
			if err != nil {
				return
			}
			fd, ok := m.FieldMap[tc.field]
			require.True(t, ok)
			assert.Equal(t, tc.wantColName, fd.ColName)
		})
	}
}

type TestModel struct {
	Id        int64
	// ""
	FirstName string
	Age       int8
	LastName  *sql.NullString
}
